major reorganization and cleaning
